Booking: refactor booking request into own service class
