+++ src/Wordpress/CustomPostType/Booking.php
@@ -6,9 +6,6 @@ use CommonsBooking\Exception\BookingDeniedException;
 use CommonsBooking\Exception\TimeframeInvalidException;
 use CommonsBooking\Helper\Helper;
 use CommonsBooking\Messages\BookingMessage;
-use CommonsBooking\Service\BookingRuleApplied;
-use CommonsBooking\Service\iCalendar;
-use Exception;
 use function wp_verify_nonce;
 
 /**
@@ -205,7 +202,7 @@ class Booking extends Timeframe {
 			$postName        = isset( $_REQUEST['cb_booking'] ) ? sanitize_text_field( wp_unslash( $_REQUEST['cb_booking'] ) ) : null;
 			$postType        = isset( $_REQUEST['type'] ) ? sanitize_text_field( wp_unslash( $_REQUEST['type'] ) ) : null;
 
-			$postId = self::handleBookingRequest(
+			$postId = \CommonsBooking\Service\Booking::handleBookingRequest(
 				$itemId,
 				$locationId,
 				$post_status,
@@ -226,255 +223,6 @@ class Booking extends Timeframe {
 		}
 	}
 
-
-	/**
-	 *
-	 * Will handle the frontend booking request. We moved this to a separate function
-	 * so that we can test it.
-	 *
-	 * @param string|null $itemId
-	 * @param string|null $locationId
-	 * @param string|null $post_status
-	 * @param int|null    $post_ID
-	 * @param string|null $comment
-	 * @param string|null $repetitionStart
-	 * @param string|null $repetitionEnd
-	 * @param string|null $requestedPostName
-	 * @param string|null $postType
-	 *
-	 * @return int - the post id of the created booking
-	 * @throws BookingDeniedException - if the booking is not possible, message contains translated text for the user
-	 */
-	public static function handleBookingRequest(
-		?string $itemId,
-		?string $locationId,
-		?string $post_status,
-		?int $post_ID,
-		?string $comment,
-		?string $repetitionStart,
-		?string $repetitionEnd,
-		?string $requestedPostName,
-		?string $postType,
-		int $overbookedDays = 0
-	): int {
-
-		if ( isset( $_POST['calendar-download'] ) ) {
-			try {
-				iCalendar::downloadICS( $post_ID );
-			} catch ( Exception $e ) {
-				// redirect to booking page and do nothing
-				return $post_ID;
-			}
-			exit;
-		}
-
-		if ( $itemId === null || ! filter_var( $itemId, FILTER_VALIDATE_INT ) || ! get_post( (int) $itemId ) ) {
-			// translators: $s = id of the item
-			throw new BookingDeniedException( sprintf( __( 'Item does not exist. (%s)', 'commonsbooking' ), $itemId ) );
-		}
-
-		if ( $locationId === null || ! filter_var( $locationId, FILTER_VALIDATE_INT ) || ! get_post( (int) $locationId ) ) {
-			// translators: $s = id of the location
-			throw new BookingDeniedException( sprintf( __( 'Location does not exist. (%s)', 'commonsbooking' ), $locationId ) );
-		}
-
-		if ( $repetitionStart === null || $repetitionEnd === null ) {
-			throw new BookingDeniedException( __( 'Start- and/or end-date is missing.', 'commonsbooking' ) );
-		}
-
-		// Validation end, set correctly typed params
-		$itemId          = (int) $itemId;
-		$locationId      = (int) $locationId;
-		$repetitionStart = (int) $repetitionStart;
-		$repetitionEnd   = (int) $repetitionEnd;
-
-		if ( $post_ID != null && ! get_post( $post_ID ) ) {
-			throw new BookingDeniedException(
-				__( 'Your reservation has expired, please try to book again', 'commonsbooking' ),
-				add_query_arg( 'cb-location', $locationId, get_permalink( get_post( $itemId ) ) )
-			);
-		}
-
-		/** @var \CommonsBooking\Model\Booking|null $booking */
-		$booking = \CommonsBooking\Repository\Booking::getByDate(
-			$repetitionStart,
-			$repetitionEnd,
-			$locationId,
-			$itemId
-		);
-
-		// Reject if the slot is already booked by another user (getExistingBookings excludes this booking by ID)
-		if ( $booking && ! commonsbooking_isCurrentUserAllowedToEdit( $booking ) ) {
-			throw new BookingDeniedException( __( 'There is already a booking in this time-range. This notice may also appear if there is an unconfirmed booking in the requested period. Unconfirmed bookings are deleted after about 10 minutes. Please try again in a few minutes.', 'commonsbooking' ) );
-		}
-
-		// The initial calendar flow always submits post_status=unconfirmed without a post_ID.
-		// If an exact-slot booking already exists for the current user, redirect to that booking
-		// instead of downgrading it back to unconfirmed via the update path below.
-		if ( $booking && $post_status === 'unconfirmed' && $post_ID === null ) {
-			return $booking->ID;
-		}
-
-		// Frontend requests should never reopen an existing booking as unconfirmed by ID.
-		// Treat this as an invalid request instead of mutating the current booking state.
-		if ( $post_status === 'unconfirmed' && $post_ID !== null ) {
-			throw new BookingDeniedException( __( 'Invalid booking request. Please try again.', 'commonsbooking' ) );
-		}
-
-		$existingBookings =
-			\CommonsBooking\Repository\Booking::getExistingBookings(
-				$itemId,
-				$locationId,
-				$repetitionStart,
-				$repetitionEnd,
-				$booking->ID ?? null,
-			);
-
-		// delete unconfirmed booking if booking process is canceled by user
-		if ( $post_status === 'delete_unconfirmed' && $booking->ID === $post_ID ) {
-			wp_delete_post( $post_ID );
-			throw new BookingDeniedException(
-				__( 'Booking canceled.', 'commonsbooking' ),
-				add_query_arg( 'cb-location', $locationId, get_permalink( get_post( $itemId ) ) )
-			);
-		}
-
-		// Validate booking -> check if there are no existing bookings in timerange.
-		if ( count( $existingBookings ) > 0 ) {
-			// checks if it's an edit, but ignores exact start/end time
-			$isEdit = count( $existingBookings ) === 1 &&
-						array_values( $existingBookings )[0]->getPost()->post_name === $requestedPostName &&
-						intval( array_values( $existingBookings )[0]->getPost()->post_author ) === get_current_user_id();
-
-			if ( ! $isEdit && $post_status !== 'canceled' ) {
-				if ( $booking ) {
-					$post_status = 'unconfirmed';
-				} else {
-					throw new BookingDeniedException( __( 'There is already a booking in this time-range. This notice may also appear if there is an unconfirmed booking in the requested period. Unconfirmed bookings are deleted after about 10 minutes. Please try again in a few minutes.', 'commonsbooking' ) );
-				}
-			}
-		}
-
-		$postarr['type']                  = $postType;
-		$postarr['post_status']           = $post_status;
-		$postarr['post_type']             = self::getPostType();
-		$postarr['post_title']            = esc_html__( 'Booking', 'commonsbooking' );
-		$postarr['meta_input']['comment'] = $comment;
-
-		// New booking
-		if ( empty( $booking ) ) {
-			if ( $post_status !== 'unconfirmed' ) {
-				// New bookings always have to be unconfirmed
-				throw new BookingDeniedException( __( 'Invalid booking request. Please try again.', 'commonsbooking' ) );
-			}
-
-			$unconfirmedBookings = array_filter(
-				\CommonsBooking\Repository\Booking::getForCurrentUser( true, null, [ 'unconfirmed' ] ),
-				fn( $booking ) => intval( $booking->getPost()->post_author ) === get_current_user_id() // getForCurrentUser also gets managed bookings, these should be ignored
-			);
-			if ( ! empty( $unconfirmedBookings ) ) {
-				if ( count( $unconfirmedBookings ) === 1 ) {
-					$bookingLink = reset( $unconfirmedBookings )->bookingLink();
-					throw new BookingDeniedException(
-						__( 'You already have an unconfirmed booking. Please confirm / cancel your existing booking or wait for it to expire.', 'commonsbooking' ) .
-						'<br>' . $bookingLink
-					);
-				} else {
-					throw new BookingDeniedException( __( 'You already have unconfirmed bookings. Please wait a few minutes for them to expire or confirm/cancel them before creating a new one.', 'commonsbooking' ) );
-				}
-			}
-			$postarr['post_name']  = Helper::generateRandomString();
-			$postarr['meta_input'] = array(
-				\CommonsBooking\Model\Timeframe::META_LOCATION_ID   => $locationId,
-				\CommonsBooking\Model\Timeframe::META_ITEM_ID       => $itemId,
-				\CommonsBooking\Model\Timeframe::REPETITION_START   => $repetitionStart,
-				\CommonsBooking\Model\Timeframe::REPETITION_END     => $repetitionEnd,
-				'type'                                              => Timeframe::BOOKING_ID,
-			);
-
-			$postId          = wp_insert_post( $postarr, true );
-			$needsValidation = true;
-
-			// Existing booking
-		} else {
-			$postarr['ID'] = $booking->ID;
-			if ( $postarr['post_status'] === 'canceled' ) {
-				$postarr['meta_input']['cancellation_time'] = current_time( 'timestamp' );
-			}
-			$postId = wp_update_post( $postarr );
-
-			// we check if this is an already denied booking and demand validation again
-			if ( $postarr['post_status'] == 'unconfirmed' ) {
-				$needsValidation = true;
-			} else {
-				$needsValidation = false;
-			}
-		}
-
-		self::saveGridSizes( $postId, $locationId, $itemId, $repetitionStart, $repetitionEnd );
-
-		$bookingModel = new \CommonsBooking\Model\Booking( $postId );
-		// we need some meta-fields from bookable-timeframe, so we assign them here to the booking-timeframe
-		try {
-			$bookingModel->assignBookableTimeframeFields();
-			if ( $overbookedDays > 0 ) { // avoid setting the value when not present (for example when updating the booking)
-				$bookingModel->setOverbookedDays( $overbookedDays );
-			}
-		} catch ( \Exception $e ) {
-			throw new BookingDeniedException(
-				__( 'There was an error while saving the booking. Please try again. Thrown error:', 'commonsbooking' ) .
-												PHP_EOL . $e->getMessage()
-			);
-		}
-
-		// check if the Booking we want to create conforms to the set booking rules
-		if ( $needsValidation ) {
-			try {
-				BookingRuleApplied::bookingConformsToRules( $bookingModel );
-			} catch ( BookingDeniedException $e ) {
-				wp_delete_post( $bookingModel->ID );
-				throw new BookingDeniedException( $e->getMessage() );
-			}
-		}
-
-		if ( $postId instanceof \WP_Error ) {
-			throw new BookingDeniedException(
-				__( 'There was an error while saving the booking. Please try again. Resulting WP_ERROR: ', 'commonsbooking' ) .
-												PHP_EOL . implode( ', ', $postId->get_error_messages() )
-			);
-		}
-
-		return $postId;
-	}
-
-	/**
-	 * Multi grid size
-	 * We need to save the grid size for timeframes with full slot grid.
-	 *
-	 * @param $postId
-	 * @param $locationId
-	 * @param $itemId
-	 * @param $startDate
-	 * @param $endDate
-	 */
-	private static function saveGridSizes( $postId, $locationId, $itemId, $startDate, $endDate ): void {
-		$startTimeFrame = \CommonsBooking\Repository\Timeframe::getByLocationItemTimestamp( $locationId, $itemId, $startDate );
-		if ( $startTimeFrame && ! $startTimeFrame->isFullDay() && $startTimeFrame->getGrid() == 0 ) {
-			update_post_meta(
-				$postId,
-				\CommonsBooking\Model\Booking::START_TIMEFRAME_GRIDSIZE,
-				$startTimeFrame->getGridSize()
-			);
-		}
-		$endTimeFrame = \CommonsBooking\Repository\Timeframe::getByLocationItemTimestamp( $locationId, $itemId, $endDate );
-		if ( $endTimeFrame && ! $endTimeFrame->isFullDay() && $endTimeFrame->getGrid() == 0 ) {
-			update_post_meta(
-				$postId,
-				\CommonsBooking\Model\Booking::END_TIMEFRAME_GRIDSIZE,
-				$endTimeFrame->getGridSize()
-			);
-		}
-	}
 
 	/**
 	 * @inheritDoc
