+++ tests/php/Wordpress/CustomPostType/BookingTest.php
@@ -27,7 +27,7 @@ class BookingTest extends CustomPostTypeTest {
 		$date->modify( '-1 day' );
 		ClockMock::freeze( $date );
 		// Case 1: We create an unconfirmed booking for a bookable timeframe. The unconfirmed booking should be created
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -50,7 +50,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertFalse( $bookingModel->isConfirmed() );
 
 		// Case 2: We now confirm the booking. The booking should be confirmed
-		$confirmedBookingId = Booking::handleBookingRequest(
+		$confirmedBookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -73,7 +73,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 3: We now try to cancel our booking a little bit later. The booking should be cancelled.
 		$date->modify( '+ 5 hours' );
 		ClockMock::freeze( $date );
-		$canceledId         = Booking::handleBookingRequest(
+		$canceledId         = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'canceled',
@@ -100,7 +100,7 @@ class BookingTest extends CustomPostTypeTest {
 	public function testHandleBookingRequest_deleteUnconfirmed() {
 		ClockMock::freeze( new \DateTime( self::CURRENT_DATE ) );
 		// We create an unconfirmed booking and then cancel the booking. The booking should be canceled
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -119,7 +119,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertTrue( $bookingModel->isUnconfirmed() );
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Booking canceled.' );
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'delete_unconfirmed',
@@ -139,7 +139,7 @@ class BookingTest extends CustomPostTypeTest {
 		$date->modify( '-1 day' );
 		ClockMock::freeze( $date );
 		// 3 Days are overbooked, that means that the Litepicker had 3 locked / holidays in range
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -163,7 +163,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertFalse( $bookingModel->isConfirmed() );
 
 		// The overbooked days are not present anymore when confirming the booking cause they are only calculated on the Litepicker screen
-		$confirmedBookingId = Booking::handleBookingRequest(
+		$confirmedBookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -194,7 +194,7 @@ class BookingTest extends CustomPostTypeTest {
 	 */
 	public function testHandleBookingRequest_noDirectCreation() {
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -206,7 +206,7 @@ class BookingTest extends CustomPostTypeTest {
 			\CommonsBooking\Wordpress\CustomPostType\Timeframe::BOOKING_ID
 		);
 		$this->bookingIds[] = $bookingId;
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -219,7 +219,7 @@ class BookingTest extends CustomPostTypeTest {
 		);
 
 		// cancel once
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'canceled',
@@ -233,7 +233,7 @@ class BookingTest extends CustomPostTypeTest {
 
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		// cancel twice, should throw exception
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'canceled',
@@ -254,7 +254,7 @@ class BookingTest extends CustomPostTypeTest {
 	 * @return void
 	 */
 	public function testHandleBookingRequest_onlyOneUnconfirmedBooking() {
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -269,7 +269,7 @@ class BookingTest extends CustomPostTypeTest {
 
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessageMatches( '/You already have an unconfirmed booking/' );
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -291,7 +291,7 @@ class BookingTest extends CustomPostTypeTest {
 	public function testHandleBookingRequest_onlyOneUnconfirmedBooking_withAdmin() {
 		$this->createSubscriber();
 		wp_set_current_user( $this->subscriberId );
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -307,7 +307,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->createAdministrator();
 		wp_set_current_user( $this->adminUserID );
 		// if this test fails, an exception would be thrown
-		$bookingTwoId       = Booking::handleBookingRequest(
+		$bookingTwoId       = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -326,7 +326,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 1: We try to create a booking without a defined location
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Location does not exist. ()' );
-		$booking = Booking::handleBookingRequest(
+		$booking = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			null,
 			'unconfirmed',
@@ -342,7 +342,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 2: We try to create a booking without a defined item
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Item does not exist. ()' );
-		$booking = Booking::handleBookingRequest(
+		$booking = \CommonsBooking\Service\Booking::handleBookingRequest(
 			null,
 			$this->locationId,
 			'unconfirmed',
@@ -358,7 +358,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 3: We try to create a booking without a defined start date
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Start- and/or end-date is missing.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -374,7 +374,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 4: We try to create a booking without a defined end date
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Start- and/or end-date is missing.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -391,7 +391,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->createConfirmedBookingStartingToday();
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'There is already a booking in this time-range. This notice may also appear if there is an unconfirmed booking in the requested period. Unconfirmed bookings are deleted after about 10 minutes. Please try again in a few minutes.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -406,7 +406,7 @@ class BookingTest extends CustomPostTypeTest {
 
 	public function testReAccessUnconfirmed() {
 		// this tests the case where the same user tries to access their unconfirmed booking again
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -423,7 +423,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertIsInt( $bookingId );
 		$bookingModel = new \CommonsBooking\Model\Booking( $bookingId );
 		$this->assertTrue( $bookingModel->isUnconfirmed() );
-		$sameBookingId      = Booking::handleBookingRequest(
+		$sameBookingId      = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -450,7 +450,7 @@ class BookingTest extends CustomPostTypeTest {
 	 * @return void
 	 */
 	public function testHandleBookingRequest_deleted_confirm() {
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -470,7 +470,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->expectExceptionMessage( 'Your reservation has expired, please try to book again' );
 
 		// now we try to confirm the booking
-		$confirmedId = Booking::handleBookingRequest(
+		$confirmedId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -496,7 +496,7 @@ class BookingTest extends CustomPostTypeTest {
 		$date->modify( '-1 day' );
 		ClockMock::freeze( $date );
 		// create regular booking through unconfirmed -> confirmed route
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -511,7 +511,7 @@ class BookingTest extends CustomPostTypeTest {
 
 		$bookingModel       = new \CommonsBooking\Model\Booking( $bookingId );
 		$postName           = $bookingModel->post_name;
-		$confirmedBookingId = Booking::handleBookingRequest(
+		$confirmedBookingId = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -525,7 +525,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->bookingIds[] = $confirmedBookingId;
 
 		// attempt to recreate the booking, should keep status as "confirmed" because it was not explicitly cancelled
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -678,7 +678,7 @@ class BookingTest extends CustomPostTypeTest {
 		$repetitionStart = strtotime( self::CURRENT_DATE );
 		$repetitionEnd   = strtotime( '+1 day', strtotime( self::CURRENT_DATE ) );
 
-		$user1BookingId     = Booking::handleBookingRequest(
+		$user1BookingId     = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -691,7 +691,7 @@ class BookingTest extends CustomPostTypeTest {
 		);
 		$this->bookingIds[] = $user1BookingId;
 		$postName           = get_post( $user1BookingId )->post_name;
-		Booking::handleBookingRequest(
+		\CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -707,7 +707,7 @@ class BookingTest extends CustomPostTypeTest {
 
 		$thrown = false;
 		try {
-			$result             = Booking::handleBookingRequest(
+			$result             = \CommonsBooking\Service\Booking::handleBookingRequest(
 				$this->itemId,
 				$this->locationId,
 				'unconfirmed',
@@ -737,7 +737,7 @@ class BookingTest extends CustomPostTypeTest {
 		$repetitionStart = strtotime( self::CURRENT_DATE );
 		$repetitionEnd   = strtotime( '+1 day', strtotime( self::CURRENT_DATE ) );
 
-		$user1BookingId     = Booking::handleBookingRequest(
+		$user1BookingId     = \CommonsBooking\Service\Booking::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -754,7 +754,7 @@ class BookingTest extends CustomPostTypeTest {
 
 		$thrown = false;
 		try {
-			$result             = Booking::handleBookingRequest(
+			$result             = \CommonsBooking\Service\Booking::handleBookingRequest(
 				$this->itemId,
 				$this->locationId,
 				'unconfirmed',
